ui: font and styling
