+++ web/app/habit/checkin/[challengeId]/page.tsx
@@ -1,8 +1,14 @@
-'use client';
-
 import CheckIn from '../components/checkin';
 
 import Header from '../../components/Header';
+import { generateMetadata } from '@/utils/generateMetadata';
+
+export const metadata = generateMetadata({
+  title: 'Check In',
+  description: 'Habit Builder for Happy Builders',
+  images: 'themes.png',
+  pathname: '',
+});
 
 export default function CheckInLayout() {
   return (
