+++ web/app/habit/components/ChallengeBox.tsx
@@ -7,14 +7,14 @@ export function ChallengeBox({ challenge }: { challenge: Challenge }) {
     <div className="wrapped m-4 w-full">
       <div className="flex w-full items-center justify-start no-underline">
         <div className="p-2 text-3xl"> {challengeToEmoji(challenge.type)} </div>
-        <div className="flex flex-col items-start justify-start p-2">
-          <p className="text-xs opacity-80">
+        <div className="text-primary flex flex-col items-start justify-start p-2">
+          <p className="text-xs font-bold opacity-80">
             {formatDuration(challenge.startTimestamp, challenge.endTimestamp)}
           </p>
           <p className="text-start text-sm font-bold">{challenge.name}</p>
           <p className="text-sm"> 5 joined </p>
         </div>
-        <div className="ml-auto p-2 text-sm">{challenge.targetNum} times</div>
+        <div className="ml-auto p-2 text-sm font-bold">{challenge.targetNum} times</div>
       </div>
     </div>
   );
