+++ web/app/habit/components/Header.tsx
@@ -7,17 +7,15 @@ import GetSingleTrait from '@/components/Nouns/GetSingleTrait';
 
 export default function Header() {
   return (
-    <main className="container mx-auto flex flex-col items-center px-8 pt-2">
+    <main className="container mx-auto flex flex-col items-center pt-2">
       <Toaster />
-      <Link href="/habit" className="p-2">
+      <Link href="/habit">
         <button type="button" aria-label="Nouns Logo">
-          <GetSingleTrait
-            properties={{ name: 'Nouns Logo', glasses: -2, width: 100, height: 40 }}
-          />
+          <GetSingleTrait properties={{ name: 'Nouns Logo', glasses: -2, width: 80, height: 32 }} />
         </button>
       </Link>
 
-      <div className="text-primary container mb-8 w-full text-center text-3xl font-bold">
+      <div className="text-primary container w-full pb-6 text-center font-londrina text-3xl font-bold">
         Alibuda Habit Builder{' '}
       </div>
     </main>
