+++ web/app/habit/list/challengeList.tsx
@@ -1,0 +1,33 @@
+'use client';
+
+import Header from '../components/Header';
+import useAllChallenges from '@/hooks/useAllChallenges';
+import Link from 'next/link';
+import { ChallengeBox } from '../components/ChallengeBox';
+
+export default function ChallengeList() {
+  const { challenges } = useAllChallenges();
+
+  return (
+    <main className="container mx-auto flex flex-col items-center px-8 pt-16 text-center">
+      <Header />
+
+      <div className="flex flex-col items-center justify-center">
+        <p className="pb-4 text-center font-londrina text-xl font-bold"> Join a Challenge Now! </p>
+
+        {/* Challenge List */}
+        <div className="pt-4">
+          {challenges.map((challenge) => (
+            <Link
+              className="w-full no-underline"
+              key={challenge.id.toString()}
+              href={`/habit/stake/${challenge.id}`}
+            >
+              <ChallengeBox challenge={challenge} />
+            </Link>
+          ))}
+        </div>
+      </div>
+    </main>
+  );
+}
