+++ web/app/layout.tsx
@@ -27,7 +27,7 @@ initAnalytics();
  */
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
-    <html lang="en" className={`${nunito.className} ${londrina.className}`}>
+    <html lang="en" className={`${nunito.className} ${londrina.variable} ${nunito.variable}`}>
       <body className="flex flex-1 flex-col">
         <OnchainProviders>
           <div className="lg:mx-auto lg:max-w-3xl lg:px-32">{children}</div>
