+++ web/package.json
@@ -119,6 +119,6 @@
   },
   "packageManager": "yarn@4.1.0",
   "resolutions": {
-    "@coinbase/wallet-sdk": "npm:@coinbase/wallet-sdk@4.0.2"
+    "@coinbase/wallet-sdk": "npm:@coinbase/wallet-sdk@4.0.0-beta.12"
   }
 }
